+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [2.0.0] WIP
  - Full reentrancy, supporting multiple concurrent evaluations of the same `Expression` object
  - New multithreading model independent of Node.js/libuv
+ - `Expression.prototype.toString()`
 
 # [1.0.0] 2021-12-26
  - Fix missing TS types